+++ packages/react-query-next-experimental/CHANGELOG.md
@@ -1,5 +1,40 @@
 # @tanstack/react-query-next-experimental
 
+## 5.100.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/react-query@5.100.4
+
+## 5.100.3
+
+### Patch Changes
+
+- Updated dependencies [[`f85d825`](https://github.com/TanStack/query/commit/f85d825e02efbbff02e2081528ed28f5e5382f7a)]:
+  - @tanstack/react-query@5.100.3
+
+## 5.100.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/react-query@5.100.2
+
+## 5.100.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/react-query@5.100.1
+
+## 5.100.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/react-query@5.100.0
+
 ## 5.99.2
 
 ### Patch Changes